+++ oop/src/PurchaseOrders/use-cases/CreatePO.usecase.ts
@@ -1,6 +1,6 @@
 import { UseCase } from "../../lib/UseCase.base";
 import { UUID, createUuid } from "../../utilities/uuid";
-import { IPORepository } from "../domain/IPORepository";
+import { IPORepository } from "../domain/PurchaseOrder.repo-interface";
 import { PurchaseOrder } from "../domain/PurchaseOrder.entity";
 import { CreatePOLineItemDTO } from "../infrastructure/dto/CreatePO.dto";
